refactor: 도메인 클래스 패키지별로 분류
- 도메인 클래스가 많아 짐에 따라 각 도메인을 유의미한 패키지별로 분류
