+++ src/main/java/domain/LadderGame.java
@@ -1,5 +1,12 @@
 package domain;
 
+import domain.ladder.Ladder;
+import domain.ladder.Line;
+import domain.player.Player;
+import domain.player.Players;
+import domain.prize.Prizes;
+import domain.prize.Results;
+
 public class LadderGame {
 
     private final Ladder ladder;
