+++ src/main/java/domain/ladder/Point.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 public enum Point {
