+++ src/main/java/domain/player/Player.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.player;
+
+import domain.MoveType;
 
 import java.util.Objects;
