+++ src/main/java/domain/prize/Prize.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.prize;
 
 import java.util.Objects;
