+++ src/test/java/domain/LineTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.ladder.Line;
+import domain.ladder.Point;
 import domain.numbergenerator.TestNumberGenerator;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.DisplayName;
