+++ src/test/java/domain/MoveTypeTest.java
@@ -1,6 +1,8 @@
 package domain;
 
+import domain.ladder.Line;
 import domain.numbergenerator.TestNumberGenerator;
+import domain.player.Player;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
