+++ src/test/java/domain/PrizeTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.prize.Prize;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
