+++ src/test/java/domain/PrizesTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.prize.Prizes;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
